requirements.txt
pyyaml>=6.0
pydantic>=2.0
click>=8.1
PyQt6>=6.6
jinja2>=3.1
weasyprint>=60.0
msgpack>=1.0.7
dnslib>=0.9.23
asyncio; python_version < "3.11"
